Add functionality to the main script as follows
Two panels are displayed, where some images appear as soon as the game
starts, using a corresponding button for it. Each round lasts 4 seconds,
where an option must be chosen from the bottom buttons row on the
player's panel. There are some counters to keep track of the number of
rounds won. Finally, the first player to reach the best of five, wins
the game. Afterwards, a restart option can be clicked